tgrambot/pyproject.toml
[project]
name = "tgrambuddy"
version = "0.1.0"
description = "Your way to manage your Telegram bot automation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.20.0.post0",
    "aiosqlite>=0.21.0",
    "alembic>=1.15.2",
    "dependency-injector>=4.46.0",
    "load-dotenv>=0.1.0",
    "opencv-python>=4.11.0.86",
    "sqlalchemy>=2.0.40",
]
